Consolidate client process management
